refactor - internal test file put in own folder and made more readable, organised
