+++ __tests__/testHelpers.ts
@@ -1,0 +1,22 @@
+// Here I will set document body's inner html to this string
+
+export const mockInnerHTML = `<ul class="sortable">
+<li class="li-first">
+  <span class="handle"></span>
+  item
+</li>
+<li>
+  <span class="another-handle"></span>
+  item 2
+</li>
+<li>
+  <span class="handle"></span>
+  item 3
+</li>
+</ul>`
+
+// // select sortable
+// ul = global.body.querySelector('.sortable')
+// li = ul.querySelector('.item-first')
+// secondLi = ul.querySelector('.item-second')
+// thirdLi = ul.querySelector('.item-second')
